chore: refactor health controller
