+++ src/routes/v1/index.ts
@@ -1,17 +1,15 @@
 import express from 'express';
-import AuthController from '../../controllers/auth';
-import { health } from '../../controllers/health';
+import AuthController from '../../controllers/v1/auth';
+import HealthController from '../../controllers/common/health';
 
 export default () => {
   const router = express.Router();
 
-  const controllers = [new AuthController()];
+  const controllers = [new HealthController(), new AuthController()];
 
   controllers.forEach((controller) => {
     router.use(controller.path, controller.router);
   });
-
-  router.use('/health', health);
 
   return router;
 };
